+++ src/util/types/expressionTree.ts
@@ -29,6 +29,7 @@ export type Expression =
   | ELowerTypeName
   | ENegateExpr
   | ENumberConstant
+  | ENullaryConstructorArgumentPattern
   | EOperator
   | EOperatorAsFunctionExpr
   | EPattern
@@ -247,6 +248,9 @@ export interface ENegateExpr extends SyntaxNode {
   nodeType: "NegateExpr";
   expression: Expression;
 }
+export interface ENullaryConstructorArgumentPattern extends SyntaxNode {
+  nodeType: "NullaryConstructorArgumentPattern";
+}
 
 export function mapSyntaxNodeToExpression(
   node: SyntaxNode | null | undefined,
@@ -293,7 +297,7 @@ export function mapSyntaxNodeToExpression(
         nodeType: "BinOpExpr",
         parts: node.children
           .map(mapSyntaxNodeToExpression)
-          .filter(Utils.notUndefined),
+          .filter(Utils.notUndefined.bind(mapSyntaxNodeToExpression)),
       } as EBinOpExpr);
     }
 
@@ -306,7 +310,7 @@ export function mapSyntaxNodeToExpression(
     case "number_constant_expr":
       return Object.assign(node, {
         nodeType: "NumberConstant",
-        isFloat: false,
+        isFloat: node.text.includes("."),
       } as ENumberConstant);
 
     case "string_constant_expr":
@@ -328,7 +332,7 @@ export function mapSyntaxNodeToExpression(
             args: node.children
               .slice(1)
               .map(mapSyntaxNodeToExpression)
-              .filter(Utils.notUndefined),
+              .filter(Utils.notUndefined.bind(mapSyntaxNodeToExpression)),
           } as EFunctionCallExpr);
         }
       }
@@ -397,7 +401,7 @@ export function mapSyntaxNodeToExpression(
         params: node.namedChildren
           .filter((n) => n.type.includes("pattern"))
           ?.map(mapSyntaxNodeToExpression)
-          .filter(Utils.notUndefined),
+          .filter(Utils.notUndefined.bind(mapSyntaxNodeToExpression)),
       } as EFunctionDeclarationLeft);
 
     case "pattern": {
@@ -425,7 +429,7 @@ export function mapSyntaxNodeToExpression(
         params: node.namedChildren
           .slice(1)
           .map(mapSyntaxNodeToExpression)
-          .filter(Utils.notUndefined),
+          .filter(Utils.notUndefined.bind(mapSyntaxNodeToExpression)),
       } as EUnionVariant);
 
     case "if_else_expr":
@@ -433,7 +437,7 @@ export function mapSyntaxNodeToExpression(
         nodeType: "IfElseExpr",
         exprList: node.namedChildren
           .map((n) => mapSyntaxNodeToExpression(n))
-          .filter(Utils.notUndefined),
+          .filter(Utils.notUndefined.bind(mapSyntaxNodeToExpression)),
       } as EIfElseExpr);
 
     case "let_in_expr":
@@ -453,7 +457,7 @@ export function mapSyntaxNodeToExpression(
         branches: node.namedChildren
           .slice(3)
           .map(mapSyntaxNodeToExpression)
-          .filter(Utils.notUndefined),
+          .filter(Utils.notUndefined.bind(mapSyntaxNodeToExpression)),
       } as ECaseOfExpr);
 
     case "case_of_branch":
@@ -490,7 +494,7 @@ export function mapSyntaxNodeToExpression(
               n.type !== "right_parenthesis",
           )
           .map(mapSyntaxNodeToExpression)
-          .filter(Utils.notUndefined),
+          .filter(Utils.notUndefined.bind(mapSyntaxNodeToExpression)),
       } as ETupleExpr);
 
     case "anything_pattern":
@@ -503,7 +507,7 @@ export function mapSyntaxNodeToExpression(
         nodeType: "TuplePattern",
         patterns: TreeUtils.findAllNamedChildrenOfType("pattern", node)
           ?.map(mapSyntaxNodeToExpression)
-          .filter(Utils.notUndefined),
+          .filter(Utils.notUndefined.bind(mapSyntaxNodeToExpression)),
       } as ETuplePattern);
 
     case "tuple_type":
@@ -514,7 +518,7 @@ export function mapSyntaxNodeToExpression(
           node,
         )
           ?.map(mapSyntaxNodeToExpression)
-          .filter(Utils.notUndefined),
+          .filter(Utils.notUndefined.bind(mapSyntaxNodeToExpression)),
         unitExpr: mapSyntaxNodeToExpression(
           TreeUtils.findFirstNamedChildOfType("unit_expr", node),
         ),
@@ -526,7 +530,7 @@ export function mapSyntaxNodeToExpression(
         exprList: node.namedChildren
           .filter((n) => n.type.endsWith("expr"))
           .map(mapSyntaxNodeToExpression)
-          .filter(Utils.notUndefined),
+          .filter(Utils.notUndefined.bind(mapSyntaxNodeToExpression)),
       } as EListExpr);
 
     case "list_pattern":
@@ -535,21 +539,24 @@ export function mapSyntaxNodeToExpression(
         parts: node.namedChildren
           .filter((n) => n.type.includes("pattern"))
           .map(mapSyntaxNodeToExpression)
-          .filter(Utils.notUndefined),
+          .filter(Utils.notUndefined.bind(mapSyntaxNodeToExpression)),
       } as EListPattern);
 
     case "union_pattern":
       return Object.assign(node, {
         nodeType: "UnionPattern",
         constructor: node.firstNamedChild,
-        namedParams: node.namedChildren
-          .filter((n) => n.type.includes("pattern"))
+        namedParams: node
+          .descendantsOfType("lower_pattern")
           .map(mapSyntaxNodeToExpression)
-          .filter(Utils.notUndefined),
+          .filter(Utils.notUndefined.bind(mapSyntaxNodeToExpression)),
         argPatterns: node.namedChildren
-          .filter((n) => n.type.includes("pattern"))
-          .map(mapSyntaxNodeToExpression)
-          .filter(Utils.notUndefined),
+          .slice(1)
+          .filter(
+            (node) =>
+              node.type.includes("pattern") || node.type.includes("constant"),
+          )
+          .map((node) => mapSyntaxNodeToExpression(node) ?? node),
       } as EUnionPattern);
 
     case "cons_pattern":
@@ -558,7 +565,7 @@ export function mapSyntaxNodeToExpression(
         parts: node.namedChildren
           .filter((n) => n.type.includes("pattern"))
           .map(mapSyntaxNodeToExpression)
-          .filter(Utils.notUndefined),
+          .filter(Utils.notUndefined.bind(mapSyntaxNodeToExpression)),
       } as EConsPattern);
     case "record_type":
       return Object.assign(node, {
@@ -568,7 +575,7 @@ export function mapSyntaxNodeToExpression(
         ),
         fieldTypes: TreeUtils.findAllNamedChildrenOfType("field_type", node)
           ?.map(mapSyntaxNodeToExpression)
-          .filter(Utils.notUndefined),
+          .filter(Utils.notUndefined.bind(mapSyntaxNodeToExpression)),
       } as ERecordType);
     case "field_type":
       return Object.assign(node, {
@@ -590,7 +597,7 @@ export function mapSyntaxNodeToExpression(
         typeVariables:
           TreeUtils.findAllNamedChildrenOfType("lower_type_name", node)
             ?.map(mapSyntaxNodeToExpression)
-            .filter(Utils.notUndefined) ?? [],
+            .filter(Utils.notUndefined.bind(mapSyntaxNodeToExpression)) ?? [],
         typeExpression: mapSyntaxNodeToExpression(
           TreeUtils.findFirstNamedChildOfType("type_expression", node),
         ),
@@ -615,7 +622,7 @@ export function mapSyntaxNodeToExpression(
         nodeType: "RecordPattern",
         patternList: TreeUtils.findAllNamedChildrenOfType("lower_pattern", node)
           ?.map(mapSyntaxNodeToExpression)
-          .filter(Utils.notUndefined),
+          .filter(Utils.notUndefined.bind(mapSyntaxNodeToExpression)),
       } as ERecordPattern);
     case "record_expr":
       return Object.assign(node, {
@@ -626,7 +633,7 @@ export function mapSyntaxNodeToExpression(
         ),
         fields: TreeUtils.findAllNamedChildrenOfType("field", node)
           ?.map(mapSyntaxNodeToExpression)
-          .filter(Utils.notUndefined),
+          .filter(Utils.notUndefined.bind(mapSyntaxNodeToExpression)),
       } as ERecordExpr);
     case "port_annotation":
       return Object.assign(node, {
@@ -657,6 +664,10 @@ export function mapSyntaxNodeToExpression(
         nodeType: "NegateExpr",
         expression: mapSyntaxNodeToExpression(node.lastNamedChild),
       } as ENegateExpr);
+    case "nullary_constructor_argument_pattern":
+      return Object.assign(node, {
+        nodeType: "NullaryConstructorArgumentPattern",
+      } as ENullaryConstructorArgumentPattern);
     default:
       return mapSyntaxNodeToExpression(node.firstNamedChild);
   }