+++ test/typeInference.test.ts
@@ -520,4 +520,58 @@ func a b =
 `;
     await testTypeInference(basicsSources + source, "Int -> Float -> Model");
   });
+
+  test("let pattern defined after used in let expr", async () => {
+    const source = `
+--@ Test.elm
+module Test exposing (..)
+
+func a b =
+--^
+  let
+    var = c
+
+    { c, d } = a
+  in
+  var
+
+`;
+    await testTypeInference(
+      basicsSources + source,
+      "{ a | c : b, d : c } -> d -> e",
+    );
+  });
+
+  test("floats", async () => {
+    const source = `
+--@ Test.elm
+module Test exposing (..)
+
+func a =
+--^
+  3.14 + a
+
+`;
+    await testTypeInference(basicsSources + source, "Float -> Float");
+  });
+
+  test("nullary constructor", async () => {
+    const source = `
+--@ Test.elm
+module Test exposing (..)
+
+type Maybe a = Just a | Nothing
+
+func =
+--^
+  case foo of
+    Just Nothing ->
+      ""
+    
+    _ ->
+      ""
+
+`;
+    await testTypeInference(basicsSources + source, "String");
+  });
 });